+++ src/pages/chat/components/search/searchItem.css
@@ -1,9 +1,8 @@
 .wrap {
     width: 30%;
     position: absolute;
-    top: 3%;
-    left: 50%;
-    transform: translate(-50%, -50%);
+    top: 10px;
+    right: 20px;
 }
 
 .search {
@@ -20,7 +19,7 @@
     border-radius: 5px;
     outline: none;
     color: #6c8078;
-    background: rgba(239, 239, 233, 0.5);
+    background: rgba(239, 239, 233, 0.7);
 }
 
 .search-text:focus {
